requirements.txt
attrs==23.1.0
bcc==0.18.0
bcrypt==4.0.1
blinker==1.7.0
Brlapi==0.8.3
cachetools==5.3.2
certifi==2020.6.20
chardet==4.0.0
chrome-gnome-shell==0.0.0
click==8.1.7
colorama==0.4.4
command-not-found==0.3
cryptography==3.4.8
cupshelpers==1.0
dbus-python==1.2.18
defer==1.0.6
distro==1.7.0
flasgger==0.9.7.1
Flask==3.0.0
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.0
Flask-Dance==7.0.0
Flask-Mail==0.9.1
Flask-SQLAlchemy==3.1.1
flask-swagger-ui==4.11.1
google-auth==2.23.4
google-auth-oauthlib==1.1.0
greenlet==3.0.1
hidpidaemon==18.4.6
httplib2==0.20.2
idna==3.3
importlib-metadata==4.6.4
itsdangerous==2.1.2
jeepney==0.7.1
Jinja2==3.1.2
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
kernelstub==3.1.4
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
louis==3.20.0
macaroonbakery==1.3.1
MarkupSafe==2.1.3
mistune==3.0.2
more-itertools==8.10.0
mysql-connector-python==8.2.0
netaddr==0.8.0
netifaces==0.11.0
oauthlib==3.2.0
packaging==23.2
pop-transition==1.1.2
protobuf==4.21.12
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycairo==1.20.1
pycups==2.0.1
pydbus==0.6.0
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
python-apt==2.4.0+ubuntu2
python-debian==0.1.43+ubuntu1.1
python-dotenv==1.0.0
python-gnupg==0.4.8
python-xlib==0.29
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
referencing==0.30.2
repolib==2.2.1
repoman==1.4.0
requests==2.25.1
requests-oauthlib==1.3.1
rpds-py==0.12.0
rsa==4.9
SecretStorage==3.3.1
sessioninstaller==0.0.0
six==1.16.0
SQLAlchemy==2.0.23
systemd-python==234
typing_extensions==4.8.0
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
urllib3==1.26.5
URLObject==2.4.3
wadllib==1.3.6
Werkzeug==3.0.1
xdg==5
xkit==0.0.0
zipp==1.0.0
